+++ src/websockets/legacy/framing.py
@@ -6,7 +6,7 @@ of frames is implemented in :mod:`websockets.legacy.protocol`.
 
 See `section 5 of RFC 6455`_.
 
-.. _section 5 of RFC 6455: http://tools.ietf.org/html/rfc6455#section-5
+.. _section 5 of RFC 6455: https://www.rfc-editor.org/rfc/rfc6455.html#section-5
 
 """