+++ src/websockets/server.py
@@ -222,7 +222,7 @@ class ServerConnection(Connection):
 
         """
         # "The user agent MUST NOT include more than one Origin header field"
-        # per https://tools.ietf.org/html/rfc6454#section-7.3.
+        # per https://www.rfc-editor.org/rfc/rfc6454.html#section-7.3.
         try:
             origin = cast(Optional[Origin], headers.get("Origin"))
         except MultipleValuesError as exc: